Fix `NameError: name 'render_templat' is not defined`.
Using `render_template` results in `TypeError: render_template() takes 1 positional argument but 2 were given`. Since there's a response object for the produced image, return that directly.  Left a TODO to handle the now orphaned stock template. It would indeed be nicer not to just return an image, but a page that contains it, among perhaps other information about the stock.